build: Add a Codecov support via Travis CI
Use docker-based provisioning to make it possible to use a proper and
up-to-date Debian base system instead of some ancient Ubuntu release.

We need to install libio-socket-ssl-perl and libmodule-install-perl
explicitly because the first has problems being built from cpanm, and
the second is missing from some of the CPAN module dependencies.